helper/compile: redid the API to be more flexible for other compiles
/cc @phinze - This will affect your WIP Ruby stuff. See the other
builtins for how to modify.

The API gets clunkier here but without this we were duplicating a lot of
types specifically for infrastructures and foundations (WIP) and I'd
rather share as much of the same compilation core as possible (such as
the customization processing).

There are aspects I like about this a lot better, too, which is there
are far fewer allocations, and far fewer intermediate types being passed
around just to shuffle data. You can just edit the data directly