core, infra/aws, helper/terraform: verify creds
This adds a new VerifyCreds step to the Infrastructure interface that
Core executes just after it fetches creds (either by asking the user or
by loading them from the cache). Infrastructure implementers can do
whatever they like in VerifyCreds in order to ensure that the
loaded credentials are good or still good. If not, a nice up-front error
message can be produced before any work is attempted.

The initial use-case for VerifyCreds is to ensure that the SSH key
provided by the user for AWS infrastructure is loaded into the local SSH
agent.

If we don't find the SSH key in the agent, we do a simple heuristic
(remove ".pub") to see if we find a private key on disk. If so, we
attempt to run ssh-add to load the key for us the user.