home: livestream when LIVESTREAM_ACTIVE
This adds some stuff to takeover the whole site with
an embedded livestream. Our press goes out at 10am with links
to this website embedded in releases, but we don't want
to have folks see it until the keynote is concluded.

Unsetting LIVESTREAM_ACTIVE and re-deploying will
open it up.

Also fixed a configuration option.