appfile: fix non-cwd compile
The `otto compile` command can take a dir argument, but Core was
assuming it could use cwd in RootDir() to look for the `.otto`
directory. This moves the wd assumption out of RootDir and up in to
Appfile(), which is used by the other non-compile commands.

Should fix the build.